Move ultratb test to be run on CI.
