+++ IPython/core/tests/test_ultratb.py
@@ -2,9 +2,13 @@
 """Tests for IPython.core.ultratb
 """
 import io
+import sys
 import os.path
 from textwrap import dedent
+import traceback
 import unittest
+
+from ..ultratb import ColorTB, VerboseTB
 
 
 from IPython.testing import tools as tt
@@ -220,3 +224,48 @@ except Exception:
             with tt.AssertNotPrints("ZeroDivisionError"), \
                     tt.AssertPrints("ValueError", suppress=False):
                 ip.run_cell(self.SUPPRESS_CHAINING_CODE)
+
+
+#----------------------------------------------------------------------------
+
+# module testing (minimal)
+def test_handlers():
+    def spam(c, d_e):
+        (d, e) = d_e
+        x = c + d
+        y = c * d
+        foo(x, y)
+
+    def foo(a, b, bar=1):
+        eggs(a, b + bar)
+
+    def eggs(f, g, z=globals()):
+        h = f + g
+        i = f - g
+        return h / i
+    
+    buff = io.StringIO()
+
+    buff.write(u'')
+    buff.write(u'*** Before ***')
+    try:
+        buff.write(spam(1, (2, 3)))
+    except:
+        traceback.print_exc(file=buff)
+
+    handler = ColorTB(ostream=buff)
+    buff.write(u'*** ColorTB ***')
+    try:
+        buff.write(spam(1, (2, 3)))
+    except:
+        handler(*sys.exc_info())
+    buff.write(u'')
+
+    handler = VerboseTB(ostream=buff)
+    buff.write(u'*** VerboseTB ***')
+    try:
+        buff.write(spam(1, (2, 3)))
+    except:
+        handler(*sys.exc_info())
+    buff.write(u'')
+
