+++ IPython/core/ultratb.py
@@ -1326,9 +1326,9 @@ class AutoFormattedTB(FormattedTB):
 class ColorTB(FormattedTB):
     """Shorthand to initialize a FormattedTB in Linux colors mode."""
 
-    def __init__(self, color_scheme='Linux', call_pdb=0):
+    def __init__(self, color_scheme='Linux', call_pdb=0, **kwargs):
         FormattedTB.__init__(self, color_scheme=color_scheme,
-                             call_pdb=call_pdb)
+                             call_pdb=call_pdb, **kwargs)
 
 
 class SyntaxTB(ListTB):
@@ -1405,47 +1405,3 @@ def eqrepr(value, repr=text_repr):
 
 def nullrepr(value, repr=text_repr):
     return ''
-
-
-#----------------------------------------------------------------------------
-
-# module testing (minimal)
-if __name__ == "__main__":
-    def spam(c, d_e):
-        (d, e) = d_e
-        x = c + d
-        y = c * d
-        foo(x, y)
-
-    def foo(a, b, bar=1):
-        eggs(a, b + bar)
-
-    def eggs(f, g, z=globals()):
-        h = f + g
-        i = f - g
-        return h / i
-
-    print('')
-    print('*** Before ***')
-    try:
-        print(spam(1, (2, 3)))
-    except:
-        traceback.print_exc()
-    print('')
-
-    handler = ColorTB()
-    print('*** ColorTB ***')
-    try:
-        print(spam(1, (2, 3)))
-    except:
-        handler(*sys.exc_info())
-    print('')
-
-    handler = VerboseTB()
-    print('*** VerboseTB ***')
-    try:
-        print(spam(1, (2, 3)))
-    except:
-        handler(*sys.exc_info())
-    print('')
-
